requirements.txt
matplotlib
snntorch